+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.1 - 2019-12-05]
+
+### Fixed
+- A regression with `--seperate-json-attributes` 
+caused by improvements in 0.6.0 to JSON parsing for non-standard xml documents.
+
 ## [0.6.0 - 2019-11-26]
 
 ### Added